Convert the repository to be a nix flake using flake-parts
I am hoping that using the nix flake standard will help me grow this
repository to contain my entire stack: both code and, eventually, NixOS
configs. I guess it can only help with any kind of collaboration.

With regards to cross compilation, to keep things simple, I did not try
to take any kind of generic approach: I am just defining cross-compiled
versions of certain targets, as I expect to only have a few
cross-compiled components (for some 32 bits Raspberry Pis).

I was able to get mostly everything to work, except for libmonome that
does not cross compile at the moment (the dontAddWafCrossFlags attribute
is not honored for some reason I do not understand), this can be
reproduced with:

    nix build '.#serialosc_docker_rpi'

I setup https://github.com/lopter/badwaf with two branches: main for a
vanilla flake, and flake-parts for a flake with flake-parts) to try
(unsuccessfully) reproduce the issue.

Now that this is public, I guess I'll have to clarify the license and
the name for the repository soon.